Cleanup servicemanagement/__init__.py by moving splitting the code into smaller files. Also split _internal.py into _common_*.py files. May end up merging some files later on, but for now it's useful to know that _common_* is code that used to be common (in azure/__init__.py) and thus may be duplicated  in other libraries (servicebus, storage).
