+++ ChangeLog.txt
@@ -1,11 +1,15 @@
 XXX-XX-XX Version X.Y.Z
 
-The following classes have moved from azure to azure.common:
+The following classes have moved from 'azure' to 'azure.common':
     WindowsAzureError
     WindowsAzureConflictError
     WindowsAzureMissingResourceError
     WindowsAzureBatchOperationError
     WindowsAzureAsyncOperationError
+
+The following have moved from 'azure' to 'azure.servicemanagement':
+    MANAGEMENT_HOST
+    DEFAULT_HTTP_TIMEOUT
 
 when we switch to requests for storage:
  - filters
