+++ azure-servicemanagement-legacy/azure/servicemanagement/_common_serialization.py
@@ -13,92 +13,16 @@
 # limitations under the License.
 #--------------------------------------------------------------------------
 import ast
-import base64
-import hashlib
-import hmac
 import sys
-import types
-import warnings
-
-if sys.version_info < (3,):
-    from cStringIO import StringIO
-    from urllib2 import quote as url_quote
-    from urllib2 import unquote as url_unquote
-    _strtype = basestring
-else:
-    from io import StringIO
-    from urllib.parse import quote as url_quote
-    from urllib.parse import unquote as url_unquote
-    _strtype = str
 
 from datetime import datetime
 from xml.sax.saxutils import escape as xml_escape
 
-from azure.common import (
-    WindowsAzureError,
-    WindowsAzureConflictError,
-    WindowsAzureMissingResourceError,
-    WindowsAzureBatchOperationError,
-    WindowsAzureAsyncOperationError,
+from ._common_models import (
+    WindowsAzureData,
+    HeaderDict,
+    _unicode_type,
 )
-
-#--------------------------------------------------------------------------
-# constants
-
-# Live ServiceClient URLs
-MANAGEMENT_HOST = 'management.core.windows.net'
-
-# Default timeout for http requests (in secs)
-DEFAULT_HTTP_TIMEOUT = 65
-
-# All of our error messages
-_ERROR_CONFLICT = 'Conflict ({0})'
-_ERROR_NOT_FOUND = 'Not found ({0})'
-_ERROR_UNKNOWN = 'Unknown error ({0})'
-_ERROR_VALUE_NONE = '{0} should not be None.'
-_ERROR_ASYNC_OP_FAILURE = 'Asynchronous operation did not succeed.'
-_ERROR_ASYNC_OP_TIMEOUT = 'Timed out waiting for async operation to complete.'
-
-METADATA_NS = 'http://schemas.microsoft.com/ado/2007/08/dataservices/metadata'
-
-
-class WindowsAzureData(object):
-
-    ''' This is the base of data class.
-    It is only used to check whether it is instance or not. '''
-    pass
-
-
-class Feed(object):
-    pass
-
-
-class _Base64String(str):
-    pass
-
-
-class HeaderDict(dict):
-
-    def __getitem__(self, index):
-        return super(HeaderDict, self).__getitem__(index.lower())
-
-
-def _encode_base64(data):
-    if isinstance(data, _unicode_type):
-        data = data.encode('utf-8')
-    encoded = base64.b64encode(data)
-    return encoded.decode('utf-8')
-
-
-def _decode_base64_to_bytes(data):
-    if isinstance(data, _unicode_type):
-        data = data.encode('utf-8')
-    return base64.b64decode(data)
-
-
-def _decode_base64_to_text(data):
-    decoded_bytes = _decode_base64_to_bytes(data)
-    return decoded_bytes.decode('utf-8')
 
 
 # TODO: check if this is used
@@ -191,45 +115,6 @@ def _get_serialization_name(element_name):
 
     return ''.join(name.capitalize() for name in element_name.split('_'))
 
-if sys.version_info < (3,):
-    _unicode_type = unicode
-
-    def _str(value):
-        if isinstance(value, unicode):
-            return value.encode('utf-8')
-
-        return str(value)
-else:
-    _str = str
-    _unicode_type = str
-
-
-def _str_or_none(value):
-    if value is None:
-        return None
-
-    return _str(value)
-
-
-def _int_or_none(value):
-    if value is None:
-        return None
-
-    return str(int(value))
-
-
-def _bool_or_none(value):
-    if value is None:
-        return None
-
-    if isinstance(value, bool):
-        if value:
-            return 'true'
-        else:
-            return 'false'
-
-    return str(value)
-
 
 # TODO: check if this is used
 def _convert_class_to_xml(source, xml_prefix=True):
@@ -268,11 +153,6 @@ def _set_continuation_from_response_headers(feeds, response):
         setattr(feeds, 'x_ms_continuation', x_ms_continuation)
 
 
-def _validate_not_none(param_name, param):
-    if param is None:
-        raise TypeError(_ERROR_VALUE_NONE.format(param_name))
-
-
 def _get_request_body(request_body):
     '''Converts an object into a request body.  If it's None
     we'll return an empty string, if it's one of our objects it'll
@@ -296,74 +176,3 @@ def _get_request_body(request_body):
         return request_body.encode('utf-8')
 
     return request_body
-
-
-class _dict_of(dict):
-
-    """a dict which carries with it the xml element names for key,val.
-    Used for deserializaion and construction of the lists"""
-
-    def __init__(self, pair_xml_element_name, key_xml_element_name,
-                 value_xml_element_name):
-        self.pair_xml_element_name = pair_xml_element_name
-        self.key_xml_element_name = key_xml_element_name
-        self.value_xml_element_name = value_xml_element_name
-        super(_dict_of, self).__init__()
-
-
-class _list_of(list):
-
-    """a list which carries with it the type that's expected to go in it.
-    Used for deserializaion and construction of the lists"""
-
-    def __init__(self, list_type, xml_element_name=None):
-        self.list_type = list_type
-        if xml_element_name is None:
-            self.xml_element_name = list_type.__name__
-        else:
-            self.xml_element_name = xml_element_name
-        super(_list_of, self).__init__()
-
-
-class _scalar_list_of(list):
-
-    """a list of scalar types which carries with it the type that's
-    expected to go in it along with its xml element name.
-    Used for deserializaion and construction of the lists"""
-
-    def __init__(self, list_type, xml_element_name):
-        self.list_type = list_type
-        self.xml_element_name = xml_element_name
-        super(_scalar_list_of, self).__init__()
-        
-class _xml_attribute:
-    
-    """a accessor to XML attributes
-    expected to go in it along with its xml element name.
-    Used for deserialization and construction"""
-    
-    def __init__(self, xml_element_name):
-        self.xml_element_name = xml_element_name
-
-
-
-
-def _general_error_handler(http_error):
-    ''' Simple error handler for azure.'''
-    if http_error.status == 409:
-        raise WindowsAzureConflictError(
-            _ERROR_CONFLICT.format(str(http_error)))
-    elif http_error.status == 404:
-        raise WindowsAzureMissingResourceError(
-            _ERROR_NOT_FOUND.format(str(http_error)))
-    else:
-        if http_error.respbody is not None:
-            raise WindowsAzureError(
-                _ERROR_UNKNOWN.format(str(http_error)) + '\n' + \
-                    http_error.respbody.decode('utf-8-sig'))
-        else:
-            raise WindowsAzureError(_ERROR_UNKNOWN.format(str(http_error)))
-
-
-def _lower(text):
-    return text.lower()
